api/requirements.txt
# python==3.12.10
fastapi==0.116.1
pydantic[email]==2.11.7
uvicorn[standard]==0.35.0
httpx==0.28.1
pytest==8.4.1
pytest-mock==3.14.1
pytest-cov==6.2.1
ruff==0.12.8
pre-commit==4.3.0
SQLAlchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.4
python-dotenv==1.1.1
